RxJava2 plugin
