+++ orbit-2-rxjava2/orbit-2-rxjava2_build.gradle.kts
@@ -14,14 +14,10 @@
  *  limitations under the License.
  */
 
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
 plugins {
     id("java-library")
     kotlin("jvm")
 }
-
-apply(from = "$rootDir/gradle/scripts/jacoco.gradle.kts")
 
 dependencies {
     implementation(kotlin("stdlib-jdk8"))
